requirements.txt
pandas
langchain
python-dotenv
langchain-community
ipykernel
bs4
requests
arxiv
PyMuPDF
langchain-Text-splitters
langchain-openai
sentence-transformers
chromadb
sentence_transformers
faiss-cpu
Ollama-API
tensorflow
langchain_chroma
langchain-openai
beautifulsoup4
streamlit
langchain_groq
fastapi
uvicorn
langserve[all]
langchain-huggingface>=0.1.0

